requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
sqlalchemy==2.0.15
python-jose==3.3.0
apscheduler==3.10.1
python-multipart==0.0.6
